application-source/pyproject.toml
[project]
name = "python-cloudhub-explorer-api"
version = "0.1.0"
description = "HD image search API built with FastAPI"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.116.1",
    "uvicorn==0.35.0",
    "sqlalchemy==2.0.44",
    "passlib[bcrypt]==1.7.4",
    "python-dotenv==1.0.1",
    "python-jose[cryptography]==3.3.0",
    "bcrypt==4.0.1",
    "google-api-python-client==2.149.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.1",
    "mega.py==1.0.8",
    "pydantic-settings==2.2.1",
    "Pillow",
    "python-multipart",
    "requests",
    "ffmpeg-python",
    "psutil",
    "tenacity>=8.5.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.4.2,<9.0.0", "pytest-asyncio>=0.23.5"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
managed = true
override-dependencies = [
    "tenacity>=8.5.0"
]

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "import-outside-toplevel",
    "no-member",
    "invalid-name",
    "self-assigning-variable",
    "protected-access",
    "redefined-outer-name",
]

[tool.isort]
known_first_party = ["app"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
ignore = ["D202", "D401"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]
